pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ts_ssl"
version = "0.1.0"
description = "Remote Sensing Time Series Contrastive Self-Supervised Learning"
dependencies = [
    "torch>=2.4",
    "scikit-learn",
    "hydra-core>=1.3.2",
    "pandas",
    "mmap-ninja",
    "lightly",
    "torchinfo",
]

[project.optional-dependencies]
gpu = [
    "pytorch-cuda>=12.4",
]
gpu_cuml = [
    "pytorch-cuda>=12.4",
    "cuml-cu12>=24.12.0",
]
full = [
    "pytorch-cuda>=12.4",
    "cuml-cu12>=24.12.0",
    "tensorboard",
    "neptune>=1.13.0",
]

[tool.setuptools]
packages = ["ts_ssl"]